feeder/requirements.txt
pydantic==2.10.6
paho-mqtt==1.5.0
amazon-orders
redis==5.2.1